TVG-103 Move search item requests into new file
